content/html/document/test/test_bug489532.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=489532
-->
<head>
  <title>Test for Bug 489532</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=489532">Mozilla Bug 489532</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script>
/** Test for Bug 489532 **/
try {
  document.createElement("<div>");
  ok(false, "Should throw.")
} catch (e) {
  ok(e instanceof DOMException, "Expected DOMException.");
  is(e.code, DOMException.INVALID_CHARACTER_ERR,
     "Expected INVALID_CHARACTER_ERR.");
}
</script>
</pre>
</body>
</html>
